feat: code refinement
