+++ src/components/Header/Header.module.scss
@@ -12,6 +12,7 @@
   .logo {
     color: #333;
     font-size: 24px;
+    line-height: 1;
   }
 
   .backButton {
@@ -19,15 +20,13 @@
     color: white;
     border-radius: 50%;
     border: none;
-    width: 25px;
-    height: 25px;
-    font-size: 20px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    line-height: 2;
+    width: 24px;
+    height: 24px;
+    font-size: 22px;
+    line-height: 1;
     padding: 0;
     margin: 0;
     text-decoration: none;
+    text-align: center;
   }
 }
