+++ src/constants/constants.ts
@@ -2,4 +2,5 @@ export default {
   SITE_NAME: "PATIENTSKY NEWS",
   NEWS_DETAILS: "NEWS DETAILS",
   DATA_FETCH_ERROR: "Failed to retrieve data. Please retry.",
+  RELATED_NEWS: "You may also like",
 } as any;
